requirements.txt
cryptography>=41.0.0
PyQt5>=5.15.0
argon2-cffi>=23.1.0
pywin32>=306; sys_platform == 'win32'
pytest>=7.4.0